master merge build fix
Former-commit-id: 18b6d435777fb2292684e0fbf603ebec835d2014 [formerly af076431c099413700561e4ac33b833e495e29da] [formerly 53e7fb916389711d999ad7c73e7271ce450a0d23 [formerly 9b2761afe35bf7cf12a44ee703473050dafec06d]]
Former-commit-id: 8cbe28d7adaa3adc94d267973d721fff80a6dd83 [formerly e507258032aea9eaed6b9a312a5c6158f9223493]
Former-commit-id: ac42165cdc77bd5f40ae4f1fc9bdc2fe6dfb109e

+++ src/pop/individual.cc
@@ -153,7 +153,7 @@ namespace FT{
                 set_p(ml->get_weights(),params.feedback);
             else
             {   // set weights to zero
-                vector<double> w(Phi.rows(), 0);                     
+                vector<float> w(Phi.rows(), 0);                     
                 set_p(w,params.feedback);
             }
             this->yhat = ml->labels_to_vector(yh);
